+++ utils/logger.py
@@ -11,7 +11,7 @@ def make_logger(name, log_level):
     formatter = logging.Formatter(
         '%(levelname)s - %(name)s - %(asctime)s - %(message)s',
         '%Y-%m-%d %H:%M:%S')
-    fh = logging.FileHandler("{}.log".format(name))
+    fh = logging.FileHandler("{}.log".format(name), mode='w')
     fh.setFormatter(formatter)
     logger.addHandler(fh)
     ch = logging.StreamHandler()