+++ modules/Models.py
@@ -1,4 +1,4 @@
-"""Module that makes a pie chart contain model information"""
+"""Module that makes a pie chart that contains model information"""
 from exif import Image
 import plotly.graph_objects as go
 
@@ -17,6 +17,8 @@ def Model_Chart(photos, log):
                     log.debug("%s has model data", each)
             except KeyError:
                 log.debug("%s has no model data ", each)
+            except AttributeError:
+                log.debug("%s has no model data ", each)
 
     freq = {}
     for item in models:
@@ -31,6 +33,7 @@ def Model_Chart(photos, log):
         labels.append(key)
         values.append(value)
 
-    fig = go.Figure(data=[go.Pie(labels=labels, values=values,
-                                 title="Model Information")])
+    fig = go.Figure(data=[go.Pie(labels=labels, values=values)])
+    fig.update_layout(title="Model Information")
+
     return fig