+++ modules/__init__.py
@@ -3,5 +3,6 @@ from .GPSCheck import GPS_Check
 from .Models import Model_Chart
 from .PhotoTable import Stats
 from .DateTime import DateTime
+from .Flash import Flash_Chart
 
-__all__ = ["GPS_Check", "Model_Chart", "Stats", "DateTime"]
+__all__ = ["GPS_Check", "Model_Chart", "Stats", "DateTime", "Flash_Chart"]